use Harmonizers
A Harmonizer is used to cluster scroll wheel events, and another one is used for the Ease() class. Alone, this gives us a polyfill for window.requestAnimationFrame(). However, it does much more.

When another project uses scroller.js, it will now be able to integrate the scroll view into its Harmonizer tree. As a result, it can avoid redundant draws when scroll events and other animations all happen at the same time.