fixed initial jump on some touchscreens
Some touchscreen devices postpone the first touch move event for a bit of time incase the user intends to tap. This was resulting in a slight jump since the touch down event happened ~30 pixels away from the first touch moved event. Now, I do not start the dragging until the first touch move event, so the jump does not occur. This seems to be what mobile browsers do for scrolling.